+++ library/src/main/java/com/uploadcare/android/library/api/RequestHelper.kt
@@ -32,7 +32,7 @@ import javax.crypto.Mac
 import javax.crypto.spec.SecretKeySpec
 
 /**
- * A helper class for doing API calls to the Uploadcare API. Supports API version 0.6.
+ * A helper class for doing API calls to the Uploadcare API. Supports API version 0.7.
  */
 @Suppress("unused")
 class RequestHelper(private val client: UploadcareClient) {
@@ -72,7 +72,7 @@ class RequestHelper(private val client: UploadcareClient) {
         val formattedDate = rfc2822(calendar.time)
 
         requestBuilder.addHeader("Content-Type", contentType ?: JSON_CONTENT_TYPE)
-        requestBuilder.addHeader("Accept", "application/vnd.uploadcare-v0.6+json")
+        requestBuilder.addHeader("Accept", "application/vnd.uploadcare-v0.7+json")
         requestBuilder.addHeader("Date", formattedDate)
         requestBuilder.addHeader("User-Agent",
                 String.format("javauploadcare-android/%s/%s", BuildConfig.VERSION_NAME,